requirements.txt
atlassian-python-api==3.41.16
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
Deprecated==1.2.14
idna==3.10
jmespath==1.0.1
oauthlib==3.2.2
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
soupsieve==2.6
urllib3==2.2.3
wrapt==1.16.0
